bitmap Images work
bitmap image views now work and images are saved with file paths returned.

It is possible to click the add image button more than three times which should throw an error. The image views are populated from top to bottom regardless of the button used to add them. This was implememented in that sake of time.